Stats Change, 30 days of code add, Showcase added
10 days of stats
* added standardDeviation
* changed the names of other files to have their respective day

30 days of coding
* added link list example

Showcase
* new folder, meant to showcase my best coding examples

+++ 10 Days of Stats/day0_MeanMedianMode.cpp
@@ -54,9 +54,7 @@ double getMedian(int n, vector<int> arr) {
 
 /*
 PURPOSE: Finds the highest frequency number within an array. If there is a tie,
-the lower value is chosen. 
-DATE: 6-7-2019 
-PARAMETERS:
+the lower value is chosen. DATE: 6-7-2019 PARAMETERS:
 * int arr; a series of integers.
 VARIABLES:
 * int mode; the highest frequency number.
@@ -65,10 +63,9 @@ VARIABLES:
 * int num; the current number being counted.
 */
 int getMode(vector<int> arr) {
-  // sorts array so the frequency for each number can be found much easier
   sort(arr.begin(), arr.end());
   
-  int mode = 0;
+  long mode = 0;
   auto count = 0;
   auto max = 0;
   int num;
@@ -101,7 +98,7 @@ INPUT:
 OUTPUT:
 * double mean; the mean of arr. one decimal point.
 * double median; the median of arr. one decimal point.
-* int mode; the mode of arr.
+* long mode; the mode of arr.
 */
 int main() {
   auto n = 0;
